Configs - Settings - Move Tiferet Error to Settings
